requirements.txt
requests==2.31.0
Flask==3.0.2
redis==5.0.3
grpcio==1.70.0
grpcio-tools==1.70.0
